Refactoring code after review.

+++ src/comment/comment.controller.ts
@@ -1,22 +1,38 @@
 import {
+    ApiBadRequestResponse,
     ApiCreatedResponse,
     ApiNotFoundResponse,
     ApiOkResponse,
     ApiParam,
     ApiTags,
 } from '@nestjs/swagger';
-import { Body, Controller, Delete, Get, Param, Post, Put, Req, Res } from '@nestjs/common';
+import {
+    BadRequestException,
+    Body,
+    Controller,
+    Delete,
+    Get,
+    Param,
+    Post,
+    Put,
+    Req,
+    Res,
+} from '@nestjs/common';
 import { CommentService } from 'src/comment/comment.service';
 import { Response } from 'express';
-import { AuthenticatedRequest } from 'src/middlewares/interfaces/auth.middleware.interfaces';
+import { AuthenticatedRequest } from 'src/auth/interfaces/auth.middleware.interfaces';
 import { CreateCommentDto } from 'src/comment/dto/create.comment.dto';
 import { UpdateCommentDto } from 'src/comment/dto/update.comment.dto';
-import { EmailService } from 'src/comment/email.service';
+import { EmailService } from 'src/email/email.service';
+import { toPresentation } from '../presentation.response';
 
 @ApiTags('Comment')
 @Controller('/api')
 export class CommentController {
-    constructor(private readonly commentService: CommentService) {}
+    constructor(
+        private readonly commentService: CommentService,
+        private readonly emailService: EmailService,
+    ) {}
 
     @Get('comment/:photoId')
     @ApiParam({ name: 'photoId' })
@@ -28,13 +44,26 @@ export class CommentController {
         @Param('photoId') photoId: number,
     ): Promise<void> {
         const comments = await this.commentService.findComments(photoId);
-        res.render('comments', {
-            title: 'Comment',
-            layout: 'comments',
-            comments: comments.map(c => ({
-                ...c,
-                isAllowEdit: Number(req.user.id) === c.userId,
-            })),
+        return toPresentation({
+            req,
+            res,
+            data: {
+                comments: comments.map(c => ({
+                    ...c,
+                    isAllowEdit: Number(req.user.id) === c.userId,
+                })),
+            },
+            render: {
+                viewName: 'comments',
+                options: {
+                    title: 'Comment',
+                    layout: 'comments',
+                    comments: comments.map(c => ({
+                        ...c,
+                        isAllowEdit: Number(req.user.id) === c.userId,
+                    })),
+                },
+            },
         });
     }
 
@@ -47,42 +76,57 @@ export class CommentController {
         @Param('photoId') photoId: number,
         @Body() createCommentDto: CreateCommentDto,
     ): Promise<void> {
-        const emailService = new EmailService();
-        const emails = await emailService.searchEmails(createCommentDto.text);
-
-        await emailService.sendEmails({
-            text: createCommentDto.text,
-            emails,
-        });
+        const emails = await this.emailService.searchEmails(createCommentDto.text);
 
         const comment = await this.commentService.replaceEmails({
             emails,
             comment: createCommentDto.text,
         });
 
-        return await this.commentService.create({
+        await this.commentService.create({
             text: comment,
             userId: req.user.id,
             photoId,
+        });
+
+        await this.emailService.sendEmails({
+            text: createCommentDto.text,
+            emails,
         });
     }
 
     @Put('comment/:commentId')
     @ApiParam({ name: 'commentId' })
     @ApiOkResponse()
-    @ApiNotFoundResponse()
+    @ApiBadRequestResponse()
     async updateComment(
         @Param('commentId') commentId: number,
+        @Req() req: AuthenticatedRequest,
         @Body() updateCommentDto: UpdateCommentDto,
     ): Promise<void> {
-        await this.commentService.update(commentId, updateCommentDto);
+        const comment = await this.commentService.findComment(commentId);
+
+        if (comment && req.user.id === comment.userId) {
+            await this.commentService.update(commentId, updateCommentDto);
+        } else {
+            throw new BadRequestException('You have no way to update the comment');
+        }
     }
 
     @Delete('comment/:commentId')
     @ApiParam({ name: 'commentId' })
     @ApiOkResponse()
-    @ApiNotFoundResponse()
-    async deleteComment(@Param('commentId') commentId: number): Promise<void> {
-        await this.commentService.delete(commentId);
+    @ApiBadRequestResponse()
+    async deleteComment(
+        @Param('commentId') commentId: number,
+        @Req() req: AuthenticatedRequest,
+    ): Promise<void> {
+        const comment = await this.commentService.findComment(commentId);
+
+        if (comment && req.user.id === comment.userId) {
+            await this.commentService.delete(commentId);
+        } else {
+            throw new BadRequestException('You have no way to delete the comment');
+        }
     }
 }
